Added navbar to replace buttons for windows
I’ve added a bootstrap navbar, with each option opening a different
window/modal. This should make it more user-friendly and compact,
meaning we can get more in a small space.

Also added a new JS function to ensure the newly opened window floats
ontop of anything else (high z-index), as previously it was displacing
already existing windows which was a nuisance