Put functions together & tested
I’ve started to put the functions in one file to make it easier and
more understandable for me to access at a later date. This way I only
have to import one file rather than many.
I’ve written a test script at the bottom to calculate the overall force
in one dimension of a rocket - using the other functions.
I plotted atmospheric drag vs height to check it was correct