Started development of computational engine
I’ve started development of a computational engine to get a hang of the
coordinate system. The coordinates use XYZ in meters, and the angles
are Euler angles in terms of Yaw, Pitch Roll.

I’ve used an object orientated approach to make it easier to use and
reference objects (such as rocket & planet) later in the code. I will
move bits of code around later to make it neater.

At the bottom (# In[59] ) I’ve written some code to run plot the path
of the rocket in 3 dimensions. I’ve been changing the angle of the
rocket to see if the coordinate system works with the euler angles.

I’ve concluded the coordinate system works fine for in atmosphere
movements