three.js Navball
I’ve started development of a Navball (FDAI) to indicate which
direction the users rocket is facing. The ball uses three.js (a JS
wrapper for WebGL). I chose this over competitors such as Flash because
it’s very lightweight and has a high compatibility with other browsers.
It’s composed of a sphere with a custom image mapped onto the surface.
The ball rotates in accordance to Euler angles and will update via AJAX.