Added Lobby
I’ve added a lobby.html page. This communicates currently connected
players and game settings to the page for the user to view via web
sockets to give real time feedback of who is connected.
I’ve used a different namespace (channel) for the web socket
communications called ‘lobbu’ to make it easier to differentiate
between this and other game web socket namespaces.

I’ve also now included index.html which will be the first page the user
sees when they connect. I have added the required menu and username
input field, which is sent to the server by POST and set as a session
variable. This will allow me to keep track of which user is which.
I have also added a function to instantiate a user object at this point
in the name of the user, which i will use later. I have added this
because i need to keep track of all my users settings separately.

Finally I have added a CONFIG.txt file to the server folder. This will
allow an admin to change essential server settings with ease before the
server is started. I have put it in a text file because it is easier
and safer to edit than the code for new users, and will be in a
familiar format.