Working Computational Engine
I now have a minimum viable solution for my computational engine, where it correctly calculates the correct equations needed for the program to run. I've made sure it has been made in a way that enables me to be able to add more functionality to the code at a later date, with the use of functions and an object oriented design.

Currently the code is independent and runs itself however from line 157 onward will be removed as the run function will be called iteratively from the server.