Linked files
resources.py is now called from app.py (run function). This will enable
the calculations to be run every time a message is received to keep the
data flow streaming to the client.

The computational(game) data is packaged as an array along with meta
data, which is used to identify the data package uniquely. This is
important for tracking packet loss at a later date, checking if a
client is still connected and timing the interval between data being
sent to the client and data being sent back from the client (ping
response time).