Numerical Figure Selector
I’ve built a small Javascript UI item that allows you to change the
value of a variable by increments of a power of 10. E.g. 10, 1, 0.1
using up and down arrows.

I have done this because it allows more accurate control of variables
such as thrust and is very easy to understand for first time users.