Update numerical figure selector
i’ve updated the NFS UI so that the javascript function is generic and
applies to any function. This is important because I don’t have to
re-write the function many times over and is reusable.
This can be seen in the make() function.

It can be altered by passing in a list of the html IDs that it should
be altering, the value to change it to and where the decimal point
should lie.