Feedback / Variable Updating
I have now added the feature to edit game variables such as rocket
angle and % thrust to control the rocket.

The data now comes in from the client on a separate namespace (web
socket channel) to avoid confusion and mixed signals between the
updating namespace and the standard namespace. This also means I don’t
have to parse for information from a very long string, and the messages
to update are only received when required, rather than every ping
meaning the program is more efficient.

In the ui.html I have added JS functions to utilise the numeric value
changer i wrote. I’ve written functions that can be used with any ui
method of changing data. They are completely generic and can be reused
to change any values as long as those values are received by the game.
I have written it as generic functions to avoid confusing and messy
code, leaving it easy to read, as well as more efficient when adding in
more functionality at a later date