Utilised Multi-Threading
The computations are now run in a completely separate thread. This is
because the previous solution was flawed. Previously the computation
was run every time a request was made by the client to the server and
the server would calculate a ping time and use that as the time
interval between calculations however I would later encounter the
problem that in a multiplayer scenario there would be multiple pings to
the server and the the calculations would be run multiple times with
the same time parameter during the time. So the calculations could be
run 4 times (0.4 seconds worth) during 0.1 seconds etc.

To fix this the calculations now runs in a separate thread to the
server, and the server fetches the results from the resources script
rather than the run() function returning data